+++ src/main/java/it/unicam/cs/ids/piattaforma_agricola_locale/service/interfaces/IVenditoreService.java
@@ -21,5 +21,5 @@ public interface IVenditoreService {
 
     List<Certificazione> getCertificazioniAzienda(Venditore venditore);
 
-    boolean rimuoviCertificazioneDaAzienda(Venditore venditore, int idCertificazione);
+    boolean rimuoviCertificazioneDaAzienda(Venditore venditore, Long idCertificazione);
 }
